# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Organizations Directory API",
        "description": "API для справочника Организаций, Зданий, Деятельности",
        "version": "1.0.0"
    },
    "paths": {
        "/api/buildings": {
            "get": {
                "tags": [
                    "Здания"
                ],
                "summary": "Список зданий",
                "operationId": "572e55bad0698264e29e3722676f8d00",
                "responses": {
                    "200": {
                        "description": "Список зданий",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Building"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/organizations/{id}": {
            "get": {
                "tags": [
                    "Организации"
                ],
                "summary": "Получить организацию по ID",
                "operationId": "11826ee077ae441c61f2bee8476e2a14",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID организации",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация об организации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Organization"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Организация не найдена"
                    }
                }
            }
        },
        "/api/organizations/building/{building_id}": {
            "get": {
                "tags": [
                    "Организации"
                ],
                "summary": "Список организаций в здании",
                "operationId": "37739fe9c59451e4ab9151b5c338c3a5",
                "parameters": [
                    {
                        "name": "building_id",
                        "in": "path",
                        "description": "ID здания",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список организаций",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Organization"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/organizations/activity/{activity_id}": {
            "get": {
                "tags": [
                    "Организации"
                ],
                "summary": "Список организаций по виду деятельности (1 уровень)",
                "operationId": "3661fc680623f4d86620514e8d360660",
                "parameters": [
                    {
                        "name": "activity_id",
                        "in": "path",
                        "description": "ID вида деятельности",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список организаций",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Organization"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/organizations/nearby": {
            "get": {
                "tags": [
                    "Организации"
                ],
                "summary": "Организации в радиусе",
                "operationId": "f48b9ede7a0b474a104c338c5bd06a4d",
                "parameters": [
                    {
                        "name": "lat",
                        "in": "query",
                        "description": "Широта",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "lng",
                        "in": "query",
                        "description": "Долгота",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "radius",
                        "in": "query",
                        "description": "Радиус в км",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список организаций",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Organization"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ошибка в параметрах"
                    }
                }
            }
        },
        "/api/organizations/search": {
            "get": {
                "tags": [
                    "Организации"
                ],
                "summary": "Поиск организации по названию",
                "operationId": "d6a844c687da3752fc5c80c956078c68",
                "parameters": [
                    {
                        "name": "query",
                        "in": "query",
                        "description": "Текст для поиска в названии",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Результаты поиска",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Organization"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/organizations/activity/{activity_id}/search": {
            "get": {
                "tags": [
                    "Организации"
                ],
                "summary": "Поиск организаций по виду деятельности с учетом вложенных до 3 уровней",
                "operationId": "2cda203d950f4d0a588d09d1c1bcaa52",
                "parameters": [
                    {
                        "name": "activity_id",
                        "in": "path",
                        "description": "ID вида деятельности",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список организаций",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Organization"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Building": {
                "title": "Building",
                "required": [
                    "id",
                    "address",
                    "latitude",
                    "longitude"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "address": {
                        "type": "string",
                        "example": "г. Москва, ул. Ленина 1, офис 3"
                    },
                    "latitude": {
                        "type": "number",
                        "format": "float",
                        "example": 55.7558
                    },
                    "longitude": {
                        "type": "number",
                        "format": "float",
                        "example": 37.6173
                    }
                },
                "type": "object"
            },
            "Activity": {
                "title": "Activity",
                "required": [
                    "id",
                    "name"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Молочная продукция"
                    },
                    "children": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Activity"
                        }
                    }
                },
                "type": "object"
            },
            "Organization": {
                "title": "Organization",
                "required": [
                    "id",
                    "name",
                    "building",
                    "activities"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "ООО Рога и Копыта"
                    },
                    "phones": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "example": "8-923-666-13-13"
                        }
                    },
                    "building": {
                        "$ref": "#/components/schemas/Building"
                    },
                    "activities": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Activity"
                        }
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Здания",
            "description": "Здания"
        },
        {
            "name": "Организации",
            "description": "Организации"
        }
    ]
}